requirement.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.2
django-autocomplete-light==3.8.2
django-ckeditor==6.0.0
django-import-export==2.5.0
django-js-asset==1.2.2
djangorestframework==3.12.4
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.10.0.0
xlrd==2.0.1
xlwt==1.3.0
